BoSSS.Foundation: converted external indices to long

+++ src/L2-foundation/BoSSS.Foundation/LECVolumeQuadrature2.cs
@@ -1019,8 +1019,8 @@ namespace BoSSS.Foundation.Quadrature.Linear {
                 if(bLinearRequired) {
                     var BlockRes = ResultsOfIntegration.ExtractSubArrayShallow(new int[] { i, 0, 0 }, new int[] { i - 1, M - 1, N - 1 });
 
-                    int m0 = (int)this.m_RowMap.GlobalUniqueCoordinateIndex(0, jCell, 0);
-                    int n0 = (int)this.m_ColMap.GlobalUniqueCoordinateIndex(0, jCell, 0);
+                    long m0 = this.m_RowMap.GlobalUniqueCoordinateIndex(0, jCell, 0);
+                    long n0 = this.m_ColMap.GlobalUniqueCoordinateIndex(0, jCell, 0);
 
                     //for (int m = 0; m < M; m++) {
                     //    for (int n = 0; n < N; n++) {
@@ -1034,7 +1034,7 @@ namespace BoSSS.Foundation.Quadrature.Linear {
                     var BlockRes = ResultsOfIntegration.ExtractSubArrayShallow(new int[] { i, 0, offset }, new int[] { i - 1, M - 1, offset - 1 });
 
 
-                    int m0 = (int)this.m_RowMap.LocalUniqueCoordinateIndex(0, jCell, 0);
+                    int m0 = this.m_RowMap.LocalUniqueCoordinateIndex(0, jCell, 0);
 
                     for(int m = 0; m < M; m++)
                         m_Vector[m0 + m] += BlockRes[m]* a;
